+++ src/Codec/Transcode.cpp
@@ -16,7 +16,7 @@
 #include "Util/uv_errno.h"
 #include "Transcode.h"
 #include "Extension/AAC.h"
-
+#include "Common/config.h"
 #define MAX_DELAY_SECOND 3
 
 using namespace std;
