+++ src/Common/MediaSource.cpp
@@ -7,15 +7,16 @@
  * LICENSE file in the root of the source tree. All contributing project authors
  * may be found in the AUTHORS file in the root of the source tree.
  */
-
+#include <mutex>
 #include "Util/util.h"
 #include "Util/NoticeCenter.h"
 #include "Network/sockutil.h"
 #include "Network/Session.h"
 #include "MediaSource.h"
 #include "Common/config.h"
+#include "Common/Parser.h"
 #include "Record/MP4Reader.h"
-
+#include "PacketCache.h"
 using namespace std;
 using namespace toolkit;
