+++ src/Extension/Factory.cpp
@@ -23,6 +23,7 @@
 #include "G711.h"
 #include "L16.h"
 #include "Common/Parser.h"
+#include "Common/config.h"
 
 using namespace std;
