+++ src/Extension/Track.h
@@ -14,7 +14,6 @@
 #include <memory>
 #include <string>
 #include "Frame.h"
-#include "Util/RingBuffer.h"
 #include "Rtsp/Rtsp.h"
 
 namespace mediakit{
