+++ src/Rtmp/RtmpMediaSource.h
@@ -21,8 +21,7 @@
 #include "RtmpDemuxer.h"
 #include "Common/config.h"
 #include "Common/MediaSource.h"
-#include "Util/util.h"
-#include "Util/logger.h"
+#include "Common/PacketCache.h"
 #include "Util/RingBuffer.h"
 #include "Util/TimeTicker.h"
 #include "Util/ResourcePool.h"
