+++ src/Rtmp/RtmpPlayer.cpp
@@ -13,6 +13,9 @@
 #include "Util/util.h"
 #include "Util/onceToken.h"
 #include "Thread/ThreadPool.h"
+#include "Common/config.h"
+#include "Common/Parser.h"
+
 using namespace toolkit;
 using namespace std;
