+++ src/Rtmp/RtmpPusher.cpp
@@ -13,6 +13,7 @@
 #include "Util/util.h"
 #include "Util/onceToken.h"
 #include "Thread/ThreadPool.h"
+#include "Common/Parser.h"
 
 using namespace std;
 using namespace toolkit;
