+++ src/Rtp/RtpCache.h
@@ -16,7 +16,7 @@
 #include "PSEncoder.h"
 #include "RawEncoder.h"
 #include "Extension/CommonRtp.h"
-#include "Common/MediaSource.h"
+#include "Common/PacketCache.h"
 namespace mediakit{
 
 class RtpCache : protected PacketCache<toolkit::Buffer> {
