+++ src/Rtsp/RtspMediaSource.h
@@ -18,8 +18,7 @@
 #include <unordered_map>
 #include "Common/config.h"
 #include "Common/MediaSource.h"
-#include "RtpCodec.h"
-#include "Util/logger.h"
+#include "Common/PacketCache.h"
 #include "Util/RingBuffer.h"
 #include "Util/TimeTicker.h"
 #include "Util/ResourcePool.h"
