+++ srt/SrtTransport.hpp
@@ -9,7 +9,7 @@
 #include "Network/Session.h"
 #include "Poller/EventPoller.h"
 #include "Poller/Timer.h"
-
+#include "Common/Stamp.h"
 #include "Common.hpp"
 #include "NackContext.hpp"
 #include "Packet.hpp"
